Fixed conflit merge errors from float->long features #125

+++ bank/src/main/java/pt/ulisboa/tecnico/softeng/bank/domain/DepositOperation.java
@@ -5,7 +5,7 @@ import java.util.Iterator;
 
 public class DepositOperation  extends Operation{
 
-    public DepositOperation(Account account1, double value) {
+    public DepositOperation(Account account1, long value) {
 
         super(account1, value);
     }
